Give the item slot runtime its own owner and transitions (#737)
First family of #737. `PlayerInventoryRuntime` exposed its slot map through
`inventory_items_mut()` and the session mutated it directly inside
`mutate_player_inventory_runtime_like_cpp` closures - the item analogue of the
`gameplay_state_mut` problem #722 closed.

The runtime receives named transitions over the map it owns, with the anchors
C++ writes through:

* the slot install in `Player::_StoreItem` (`m_items[slot] = pItem`,
  Player.cpp:11277) and `Player::VisualizeItem` (:11535), returning whatever
  occupied the slot as the C++ callers check before overwriting
* the removal in `Player::RemoveItem` (Player.h:1412, Player.cpp:11553, nulling
  at :11612 and :11754), which its own comment describes as taking the item out
  of storage without changing it - the record is handed back unmodified
* the teardown and initialization that empty every slot (`delete m_items[i]`,
  Player.cpp:360; `Player::Create`, :403-404)

The C++ installs also set the inv-slot field, container, owner, slot and
`ITEM_CHANGED` state on the item itself. Those participants stay with their
existing owners; these methods own only the slot map, which is what the session
was reaching into.

One departure is named in the module: C++ keeps an `Item*` per slot and reads
entry and inventory type off the item, so it has no second copy to refresh.
RustyCore's slot record carries them, and the guarded update is how they stay in
step - the guid check keeps a stale caller from relabelling whatever now
occupies the slot.

The struct and its impls also leave the Player root for their own file, so the
physical ceiling on player/mod.rs is tightened 4904 -> 4802 instead of raised.
The two remaining `inventory_items_mut` uses in the session are `#[cfg(test)]`
fixture builds.

No new field, mirror or lock; behavior is unchanged.

cargo test -p wow-world --lib: 3826 passed, 0 failed, 1 ignored.
cargo test -p wow-entities --lib: 725 passed, 0 failed.
